Moved classes from Contracts to ABus as they are implementations
